[feature] enforce token base authentication for  put and delete requests

+++ app/__init__.py
@@ -80,38 +80,51 @@ def create_app(config_name):
     @app.route('/bucketlists/<int:id>', methods=['GET', 'PUT', 'DELETE'])
     def bucketlist_manipulation(id, **kwargs):
 
-        bucketlist = Bucketlist.query.filter_by(id=id).first()
-        if not bucketlist:
-            # Raise an HTTPException with a 404 not found status code
-            abort(404)
+        auth_header = request.headers.get('Authorization')
+        access_token = auth_header.split(" ")[1]
 
-        if request.method == "DELETE":
-            bucketlist.delete()
-            return {
-                "message": "bucketlist {} deleted ".format(bucketlist.id)
-            }, 200
-        elif request.method == 'PUT':
-            name = str(request.data.get('name', ''))
-            bucketlist.name = name
-            bucketlist.save()
-            response = jsonify({
-                'id': bucketlist.id,
-                'name': bucketlist.name,
-                'date_created': bucketlist.date_created,
-                'date_modified': bucketlist.date_modified
-            })
-            response.status_code = 200
-            return response
-        else:
-            # GET
-            response = jsonify({
-                'id': bucketlist.id,
-                'name': bucketlist.name,
-                'date_created': bucketlist.date_created,
-                'date_modified': bucketlist.date_modified
-            })
-            response.status_code = 200
-            return response
+        if access_token:
+            user_id = User.decode_token(access_token)
+            if not isinstance(user_id, str):
+                bucketlist = Bucketlist.query.filter_by(id=id).first()
+                if not bucketlist:
+                    # Raise an HTTPException with a 404 not found status code
+                    abort(404)
+
+                if request.method == "DELETE":
+                    bucketlist.delete()
+                    return {
+                        "message": "bucketlist {} deleted".format(bucketlist.id)
+                    }, 200
+                elif request.method == 'PUT':
+                    name = str(request.data.get('name', ''))
+                    bucketlist.name = name
+                    bucketlist.save()
+                    response = {
+                        'id': bucketlist.id,
+                        'name': bucketlist.name,
+                        'date_created': bucketlist.date_created,
+                        'date_modified': bucketlist.date_modified,
+                        'created_by': bucketlist.created_by
+                    }
+                    return make_response(jsonify(response)), 200
+                else:
+                    # GET
+                    response = jsonify({
+                        'id': bucketlist.id,
+                        'name': bucketlist.name,
+                        'date_created': bucketlist.date_created,
+                        'date_modified': bucketlist.date_modified,
+                        'created_by': bucketlist.created_by
+                    })
+                    return make_response(response), 200
+            else:
+                # user is not legit, so the payload is an error message
+                message = user_id
+                response = {
+                    'message': message
+                }
+                return make_response(jsonify(response)), 401
 
     from .auth import auth_blueprint
     app.register_blueprint(auth_blueprint)
